+++ graphql/resolvers/index.js
@@ -2,12 +2,14 @@ const userResolver = require('./userResolver');
 const articleResolver = require('./articleResolver')
 const tagResolver = require('./tagResolver')
 const versionResolver = require('./versionResolver')
+const authResolver = require('./authResolver')
 
 const rootResolver = {
   ...userResolver,
   ...articleResolver,
   ...tagResolver,
-  ...versionResolver
+  ...versionResolver,
+  ...authResolver
 };
 
 module.exports = rootResolver
