+++ graphql/resolvers/userResolver.js
@@ -1,6 +1,5 @@
 const bcrypt = require('bcryptjs');
 const Isemail = require('isemail');
-const jwt = require('jsonwebtoken');
 
 const User = require('../models/user');
 const Password = require('../models/user_password');
@@ -103,67 +102,5 @@ module.exports = {
     catch(err){
       throw err
     }
-  },
-  login: async (args, {req,res}) => {
-    try{
-      //login via email or username
-      let findProp = args.username? {username:args.username}:{email:args.email}
-      const fetchedPassword = await Password.findOne(findProp).populate("users")
-      if(!fetchedPassword){throw new Error("Password not found")}
-      if(!await bcrypt.compare(args.password,fetchedPassword.password)){
-        throw new Error("Password is incorrect");
-      }
-      const payload = {
-        usersIds:fetchedPassword.users.map(user => user._id.toString()),
-        passwordId:fetchedPassword.id,
-        admin:fetchedPassword.users.filter(user => user.admin).length > 0 ? true : false,
-        session:true
-      }
-
-      const token = jwt.sign(
-        payload,
-        process.env.JWT_SECRET_SESSION
-      )
-      const tokenCookie = jwt.sign(
-        payload,
-        process.env.JWT_SECRET_SESSION_COOKIE
-      )
-
-      //All query are async, can't channel req to another resolver
-      //req.user = payload;
-      console.log("setting cookie:",tokenCookie);
-      res.cookie("graphQL-jwt",tokenCookie,{expires:0,httpOnly:true,secure:process.env.HTTPS})
-
-      return {
-        token:token,
-        password:populatePassword(fetchedPassword),
-        users:fetchedPassword.users.map(populateUser)
-      }
-    }
-    catch(err){
-      throw err
-    }
-  },
-  refreshToken: async (_, {req}) => {
-    if(!req.user_noCSRF){
-      throw new Error("Can't refresh user without cookie");
-    }
-    const fetchedPassword = await Password.findOne({_id : req.user_noCSRF.passwordId}).populate("users")
-    if(!fetchedPassword){throw new Error("Password not found")}
-    const payload = {
-      usersIds:fetchedPassword.users.map(user => user._id.toString()),
-      passwordId:fetchedPassword.id,
-      admin:fetchedPassword.users.filter(user => user.admin).length > 0 ? true : false,
-      session:true
-    }
-    const token = jwt.sign(
-      payload,
-      process.env.JWT_SECRET_SESSION
-    )
-    return {
-      token:token,
-      password:populatePassword(fetchedPassword),
-      users:fetchedPassword.users.map(populateUser)
-    }
-  }
+  }  
 }
